united both chunkers in one module
